+++ routes.php
@@ -2,17 +2,19 @@
 //registramos las rutas en el router
 $router->get('/', 'HomeController@index');
 $router->get('/trastos', "TrastoController@index");
-$router->get('/trasto/create', "TrastoController@create");
-$router->get('/trasto/edit/{id}', "TrastoController@edit");
+$router->get('/trasto/create', "TrastoController@create", ['auth']);
+$router->get('/trasto/edit/{id}', "TrastoController@edit", ['auth']);
 $router->get('/trasto/{id}', "TrastoController@show");
 
-$router->post('/trasto', "TrastoController@store");
-$router->put('/trasto/{id}', "TrastoController@update");
-$router->delete('/trasto/{id}', "TrastoController@delete");
+$router->post('/trasto', "TrastoController@store", ['auth']);
+$router->put('/trasto/{id}', "TrastoController@update", ['auth']);
+$router->delete('/trasto/{id}', "TrastoController@delete", ['auth']);
 
-
+    
 //Rutas para la autenticación de los usuarios
-$router->get('/auth/register', 'UserController@create');
-$router->get('/auth/login', 'UserController@login');
+$router->get('/auth/register', 'UserController@create',['guest']);
+$router->get('/auth/login', 'UserController@login', ['guest']);
 
-$router->post('/auth/create', 'UserController@store');
+$router->post('/auth/create', 'UserController@store', ['guest']);
+$router->post('/auth/logout', 'UserController@logout', ['auth']);
+$router->post('/auth/login', 'UserController@authenticate', ['guest']);
